Revert "Refactor to not use file extension"
This reverts commit 7f5d33a982c3558ab347e1567eead79c6174eb70.
