--- src/scripts/common/index.js
@@ -1,5 +1,0 @@
-export * from './api';
-export * from './authentication';
-export * from './ElementHelper';
-export * from './renderPost';
-export * from './storage';
