+++ src/scripts/profile.js
@@ -1,4 +1,4 @@
-import { authentication } from './common';
+import { authentication } from './common/authentication';
 
 // Authentication check
 const user = await authentication.check();
